SongBloom/requirements.txt
huggingface-hub==0.24.6
jieba-fast==0.53
pypinyin==0.51.0
cn2an==0.5.22
wordsegment==1.3.1
g2p-en==2.1.0
lightning==2.2.1
nltk==3.8.1
omegaconf==2.2.0
torch==2.2.0
torchaudio==2.2.0
transformers==4.44.1
einops==0.8.0
spacy==3.7.4
num2words==0.5.13
descript-audio-codec==1.0.0
vector_quantize_pytorch==1.14.8